change return type of stripepochtime

+++ mttime.go
@@ -105,6 +105,6 @@ func TrimEpochDate(epoch int64) int {
 	return int(epoch) % 86400
 }
 
-func StripEpochTime(epoch int64) int {
-	return epoch - int(epoch) % 86400
+func StripEpochTime(epoch int64) int64 {
+	return epoch - epoch%86400
 }
